Include reviewers that can't vote on any label
When a user is added as a reviewer on a change an arbitrary label
is chosen and a value of 0 is instead into PatchSetApprovals. If
the user was not permitted to place any score on the change the
server dropped them from the label information, preventing display
in the ApprovalTable of the web UI.

This was missed up until now because most servers grant something
like Code-Review -1..+1 to Registered Users, so everyone can always
vote on a change. However if this grant was not given reviewers
did not appear to be added to the change.

The rewrite of the method is easier to follow. It now proceeds
through the table in a row-wise fashion, ensuring every column is
filled out. This is backwards from the way the JSON is organized,
which uses a column-first ordering. Column-first in this particular
function is difficult because the server must ensure every row
has all columns populated, and the column matrix is what is sparse
for a user. Column-first also requires rebuilding or caching the
ChangeControl for each reviewer, row-first permits this to be
trivially retained in a local variable.

+++ gerrit-server/src/main/java/com/google/gerrit/server/project/ChangeControl.java
@@ -265,7 +265,7 @@ public class ChangeControl {
     return canRemoveReviewer(approval.getAccountId(), approval.getValue());
   }
 
-  public boolean canRemoveReviewer(Account.Id reviewer, short value) {
+  public boolean canRemoveReviewer(Account.Id reviewer, int value) {
     if (getChange().getStatus().isOpen()) {
       // A user can always remove themselves.
       //
